# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VAANI AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProcessCallBody {
  audioBase64: string;
  mimeType: string;
  step?: string;
  sessionId: string;
}

export interface CallResult {
  response: string;
  transcript: string;
  /** English translation of transcript if not in English */
  translation: string;
  language: string;
  /** Detected regional dialect (e.g. Dharwad Kannada, Mysuru Kannada) */
  dialect: string;
  emotion: string;
  urgency: string;
  emergency_keywords: string[];
  issue: string;
  intent: string;
  confidence: number;
  level: number;
  /** 1=AI Full Auto, 2=AI+Agent, 3=Full Human Takeover */
  mode: number;
  /** 3 suggested responses for Mode 2 agent assist */
  suggested_responses: string[];
  /** Brief for agent during Mode 3 handover */
  agent_briefing: string;
}

export interface FeedbackBody {
  transcript: string;
  issue: string;
  emotion: string;
  language: string;
  urgency?: string;
  confidence?: number;
  level?: number;
  feedback: string;
}

export interface CallRecord {
  id: number;
  transcript: string;
  issue: string;
  emotion: string;
  language: string;
  /** @nullable */
  urgency: string | null;
  /** @nullable */
  confidence: number | null;
  /** @nullable */
  level: number | null;
  /** @nullable */
  feedback: string | null;
  createdAt: string;
}

export type AnalyticsSummaryLanguageBreakdown = { [key: string]: number };

export interface AnalyticsSummary {
  totalCalls: number;
  correctCount: number;
  wrongCount: number;
  accuracyPercent: number;
  distressedCount: number;
  calmCount: number;
  languageBreakdown: AnalyticsSummaryLanguageBreakdown;
  commonIssues: string[];
  level1Count: number;
  level2Count: number;
  level3Count: number;
}

export interface ResetSessionBody {
  sessionId: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ApiError {
  error: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GenerateGeminiImageBody {
  prompt: string;
}

export interface GenerateGeminiImageResponse {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}