# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VAANI AI Voice Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vaani/process:
    post:
      operationId: processVaaniCall
      tags: [vaani]
      summary: Process an audio call with Gemini AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessCallBody"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /vaani/feedback:
    post:
      operationId: submitVaaniFeedback
      tags: [vaani]
      summary: Submit feedback for a call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackBody"
      responses:
        "200":
          description: Feedback saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /vaani/reset-session:
    post:
      operationId: resetVaaniSession
      tags: [vaani]
      summary: Reset a session's conversation memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetSessionBody"
      responses:
        "200":
          description: Session reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ProcessCallBody:
      type: object
      properties:
        audioBase64:
          type: string
        mimeType:
          type: string
        step:
          type: string
        sessionId:
          type: string
      required:
        - audioBase64
        - mimeType
        - sessionId
    CallResult:
      type: object
      properties:
        response:
          type: string
        transcript:
          type: string
        translation:
          type: string
          description: English translation of transcript if not in English
        language:
          type: string
        dialect:
          type: string
          description: Detected regional dialect (e.g. Dharwad Kannada, Mysuru Kannada)
        emotion:
          type: string
        urgency:
          type: string
        emergency_keywords:
          type: array
          items:
            type: string
        issue:
          type: string
        intent:
          type: string
        confidence:
          type: integer
        level:
          type: integer
        mode:
          type: integer
          description: 1=AI Full Auto, 2=AI+Agent, 3=Full Human Takeover
        suggested_responses:
          type: array
          items:
            type: string
          description: 3 suggested responses for Mode 2 agent assist
        agent_briefing:
          type: string
          description: Brief for agent during Mode 3 handover
      required:
        - response
        - transcript
        - translation
        - language
        - dialect
        - emotion
        - urgency
        - emergency_keywords
        - issue
        - intent
        - confidence
        - level
        - mode
        - suggested_responses
        - agent_briefing
    FeedbackBody:
      type: object
      properties:
        transcript:
          type: string
        issue:
          type: string
        emotion:
          type: string
        language:
          type: string
        urgency:
          type: string
        confidence:
          type: integer
        level:
          type: integer
        feedback:
          type: string
      required:
        - transcript
        - issue
        - emotion
        - language
        - feedback
...
    ResetSessionBody:
      type: object
      properties:
        sessionId:
          type: string
      required:
        - sessionId
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...