# lib/api-zod/src/generated/types/callResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VAANI AI Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CallResult {
  response: string;
  transcript: string;
  /** English translation of transcript if not in English */
  translation: string;
  language: string;
  /** Detected regional dialect (e.g. Dharwad Kannada, Mysuru Kannada) */
  dialect: string;
  emotion: string;
  urgency: string;
  emergency_keywords: string[];
  issue: string;
  intent: string;
  confidence: number;
  level: number;
  /** 1=AI Full Auto, 2=AI+Agent, 3=Full Human Takeover */
  mode: number;
  /** 3 suggested responses for Mode 2 agent assist */
  suggested_responses: string[];
  /** Brief for agent during Mode 3 handover */
  agent_briefing: string;
}